Fix docker override file

+++ src/Composer/DockerPlugin.php
@@ -97,11 +97,47 @@ class DockerPlugin implements PluginInterface, EventSubscriberInterface
 
         (new Filesystem())->mirror($installDir, $appDir);
 
-        $this->updateEnvFile($appDir . '/.env');
-        $this->updateEnvLocalFile($appDir . '/.env.local');
+        $this->updateFile($appDir . '/.env', [
+            'APP_NAME=__dwsmyapp__',
+            'APP_ENV=prod',
+            'APP_INSTANCE=live',
+            'APP_DEBUG=0',
+            'APP_SECRET=<secret>',
+            'APP_PORT=8000',
+            'APP_UID=1000',
+            'APP_GID=33',
+            'APP_USER=www-data',
+            'APP_DUMP_HOURLY=0',
+            'TZ=Europe/Paris',
+            'MYSQL_ROOT_PASSWORD=<secret>',
+            'MYSQL_USER=__dwsmyapp__',
+            'MYSQL_PASSWORD=<secret>',
+            'MYSQL_DATABASE=__dwsmyapp__',
+            '#POSTGRES_USER=__dwsmyapp__',
+            '#POSTGRES_PASSWORD=<secret>',
+            '#POSTGRES_DB=__dwsmyapp__',
+            'BLACKFIRE_ENABLE=0',
+            'BLACKFIRE_LOG_LEVEL=4',
+            'BLACKFIRE_CLIENT_ID=',
+            'BLACKFIRE_CLIENT_TOKEN=',
+            'BLACKFIRE_SERVER_ID=',
+            'BLACKFIRE_SERVER_TOKEN=',
+        ]);
+
+        $this->updateFile($appDir . '/.env.local', [
+            'APP_ENV=dev',
+            'APP_INSTANCE=local',
+            'APP_DEBUG=1',
+            'APP_PORT=' . $this->config['APP_PORT_LOCAL'],
+        ]);
+
+        $this->updateFile($appDir . '/.gitignore', [
+            'docker-compose.override.yml',
+            'docker-compose.override.yaml',
+        ]);
 
         $this->searchAndReplace(
-            $appDir . '/docker-compose.local.yml',
+            $appDir . '/docker-compose.override.yaml',
             '33306:3306',
             $this->config['MYSQL_PORT_LOCAL'] . ':3306'
         );
@@ -110,6 +146,12 @@ class DockerPlugin implements PluginInterface, EventSubscriberInterface
             $appDir . '/config/nginx/proxies/local.conf',
             '127.0.0.1:8000',
             '127.0.0.1:' . $this->config['APP_PORT_LOCAL']
+        );
+
+        $this->searchAndReplace(
+            $appDir . '/README.md',
+            '{custom_local_port}',
+            $this->config['APP_PORT_LOCAL']
         );
 
         $this->io->write('  - Installing recipe <fg=green>docker-compose</fg=green> ');
@@ -124,74 +166,34 @@ class DockerPlugin implements PluginInterface, EventSubscriberInterface
      */
     private function searchAndReplace(string $file, string $search, string $replace): void
     {
+        if (!file_exists($file)) {
+            return;
+        }
+
         file_put_contents($file, str_replace($search, $replace, file_get_contents($file)));
     }
 
     /**
-     * updateEnvFile
+     * updateFile
      *
      * @param string $file
-     *
-     * @throws \Exception
+     * @param array  $items
      */
-    private function updateEnvFile(string $file): void
+    private function updateFile(string $file, array $items): void
     {
         $data = file_exists($file) ? file_get_contents($file) : '';
-        if (!preg_match('/###> phpguild\/docker-web-standard ###/', $data)) {
-            $data .=
-                PHP_EOL .
-                '###> phpguild/docker-web-standard ###' . PHP_EOL .
-                'APP_NAME=__dwsmyapp__' . PHP_EOL .
-                'APP_ENV=prod' . PHP_EOL .
-                'APP_DEBUG=0' . PHP_EOL .
-                'APP_SECRET=<secret>' . PHP_EOL .
-                'APP_PORT=8000' . PHP_EOL .
-                'APP_INSTANCE=live' . PHP_EOL .
-                'APP_UID=1000' . PHP_EOL .
-                'APP_GID=33' . PHP_EOL .
-                'APP_USER=www-data' . PHP_EOL .
-                'APP_DUMP_HOURLY=0' . PHP_EOL .
-                'TZ=Europe/Paris' . PHP_EOL .
-                'MYSQL_ROOT_PASSWORD=<secret>' . PHP_EOL .
-                'MYSQL_USER=__dwsmyapp__' . PHP_EOL .
-                'MYSQL_PASSWORD=<secret>' . PHP_EOL .
-                'MYSQL_DATABASE=__dwsmyapp__' . PHP_EOL .
-                '#POSTGRES_USER=__dwsmyapp__' . PHP_EOL .
-                '#POSTGRES_PASSWORD=<secret>' . PHP_EOL .
-                '#POSTGRES_DB=__dwsmyapp__' . PHP_EOL .
-                'BLACKFIRE_ENABLE=0' . PHP_EOL .
-                'BLACKFIRE_LOG_LEVEL=4' . PHP_EOL .
-                'BLACKFIRE_CLIENT_ID=' . PHP_EOL .
-                'BLACKFIRE_CLIENT_TOKEN=' . PHP_EOL .
-                'BLACKFIRE_SERVER_ID=' . PHP_EOL .
-                'BLACKFIRE_SERVER_TOKEN=' . PHP_EOL .
-                '###< phpguild/docker-web-standard ###' . PHP_EOL
-            ;
-            file_put_contents($file, $data);
-        }
-    }
 
-    /**
-     * updateEnvLocalFile
-     *
-     * @param string $file
-     *
-     * @throws \Exception
-     */
-    private function updateEnvLocalFile(string $file): void
-    {
-        $data = file_exists($file) ? file_get_contents($file) : '';
-        if (!preg_match('/###> phpguild\/docker-web-standard ###/', $data)) {
-            $data .=
-                PHP_EOL .
-                '###> phpguild/docker-web-standard ###' . PHP_EOL .
-                'APP_ENV=dev' . PHP_EOL .
-                'APP_DEBUG=1' . PHP_EOL .
-                'APP_PORT=' . $this->config['APP_PORT_LOCAL'] . PHP_EOL .
-                'APP_INSTANCE=local' . PHP_EOL .
-                '###< phpguild/docker-web-standard ###' . PHP_EOL
-            ;
-            file_put_contents($file, $data);
+        if (preg_match('/###> phpguild\/docker-web-standard ###/', $data)) {
+            return;
         }
+
+        $data .=
+            PHP_EOL .
+            '###> phpguild/docker-web-standard ###' . PHP_EOL .
+            implode(PHP_EOL, $items) .
+            '###< phpguild/docker-web-standard ###' . PHP_EOL
+        ;
+
+        file_put_contents($file, $data);
     }
 }
